Refactor: MainController 에서 가지고 있던 다른 로직 분리
