+++ src/main/java/me/synology/hajubal/coins/controller/MainController.java
@@ -1,44 +1,21 @@
 package me.synology.hajubal.coins.controller;
 
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import me.synology.hajubal.coins.controller.dto.CookieDto;
-import me.synology.hajubal.coins.entity.*;
-import me.synology.hajubal.coins.respository.*;
-import me.synology.hajubal.coins.schedule.Schedulers;
-import me.synology.hajubal.coins.service.UserCookieService;
-import org.springframework.beans.factory.annotation.Autowired;
+import me.synology.hajubal.coins.entity.SavedPoint;
+import me.synology.hajubal.coins.respository.SavedPointRepository;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.validation.BindingResult;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
 
 import java.util.List;
 
+@RequiredArgsConstructor
 @Slf4j
 @Controller
 public class MainController {
 
-    @Autowired
-    private UserCookieService userCookieService;
-
-    @Autowired
-    private UserCookieRepository userCookieRepository;
-
-    @Autowired
-    private SiteRepository siteRepository;
-
-    @Autowired
-    private PointUrlRepository pointUrlRepository;
-
-    @Autowired
-    private PointUrlCallLogRepository pointUrlCallLogRepository;
-
-    @Autowired
-    private Schedulers schedulers;
-
-    @Autowired
-    private SavedPointRepository savedPointRepository;
+    private final SavedPointRepository savedPointRepository;
 
     @GetMapping("/")
     public String index() {
@@ -53,113 +30,5 @@ public class MainController {
         model.addAttribute("items", all);
 
         return "dashboard";
-    }
-
-    @GetMapping("/users")
-    public String userCookies(Model model) {
-        List<UserCookie> list = userCookieRepository.findAll();
-
-        model.addAttribute("items", list);
-
-        return "users";
-    }
-
-    @GetMapping("/sites")
-    public String sites(Model model) {
-        List<Site> list = siteRepository.findAll();
-
-        model.addAttribute("items", list);
-
-        return "sites";
-    }
-
-    @GetMapping("/pointurl")
-    public String pointurl(Model model) {
-        List<PointUrl> list = pointUrlRepository.findAll();
-
-        model.addAttribute("items", list);
-
-        return "pointurl";
-    }
-
-    @GetMapping("/user/{userId}")
-    public String updateUser(@PathVariable Long userId, Model model) {
-
-        UserCookie userCookie = userCookieRepository.findById(userId).orElseThrow(() -> new IllegalArgumentException("Not found user."));
-
-        model.addAttribute("userCookie", userCookie);
-
-        return "editUser";
-    }
-
-    @PostMapping("/user/{userId}")
-    public String updateUser(@PathVariable Long userId, @Validated @ModelAttribute("userCookie") CookieDto.CookieUpdateDto cookieUpdateDto
-            , BindingResult bindingResult) {
-        if (bindingResult.hasErrors()) {
-            log.info("errors={}", bindingResult);
-            return "editUser";
-        }
-
-        userCookieService.updateUserCookie(userId, cookieUpdateDto);
-
-        return "redirect:/user/" + userId;
-    }
-    
-    @GetMapping("/insertUser")
-    public String insertUser(Model model) {
-        model.addAttribute("userCookie", new CookieDto.CookieInsertDto());
-
-        return "addUser";
-    }
-
-    @PostMapping("/insertUser")
-    public String insertCookie(@Validated @ModelAttribute("userCookie") CookieDto.CookieInsertDto cookieInsertDto, BindingResult bindingResult) {
-        if (bindingResult.hasErrors()) {
-            log.info("errors={}", bindingResult);
-            return "addUser";
-        }
-
-        Long userId = userCookieService.insertUserCookie(cookieInsertDto);
-
-        return "redirect:/user/" + userId;
-    }
-
-    @DeleteMapping("/user/{userId}")
-    public String deleteUser(@PathVariable Long userId) {
-
-        userCookieRepository.deleteById(userId);
-
-        return "redirect:/users";
-    }
-
-    @GetMapping("/crawling")
-    public String crawling() {
-        schedulers.webCrawlerScheduler();
-
-        return "redirect:/pointurl";
-    }
-
-    @GetMapping("/savePoint")
-    public String savePoint() {
-        schedulers.pointScheduler();
-
-        return "redirect:/pointurl";
-    }
-
-    @GetMapping("/savePointLog")
-    public String savePointLog(Model model, @RequestParam(name = "userName", defaultValue = "ha") String[] userName) {
-
-        List<PointUrlCallLog> list;
-
-        if (userName.length == 0) {
-            list = pointUrlCallLogRepository.findAll();
-        } else {
-            list = pointUrlCallLogRepository.findByUserNameIn(userName);
-        }
-
-        model.addAttribute("userName", userName);
-        model.addAttribute("items", list);
-
-        return "pointLog";
     }
 }
