+++ src/main/java/me/synology/hajubal/coins/service/SiteService.java
@@ -1,0 +1,23 @@
+package me.synology.hajubal.coins.service;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import me.synology.hajubal.coins.entity.Site;
+import me.synology.hajubal.coins.respository.SiteRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@RequiredArgsConstructor
+@Slf4j
+@Transactional(readOnly = true)
+@Service
+public class SiteService {
+
+    private final SiteRepository siteRepository;
+
+    public List<Site> getAll() {
+        return siteRepository.findAll();
+    }
+}
