+++ src/main/resources/templates/layout/main.html
@@ -60,7 +60,7 @@
             </a>
           </li>
           <li class="nav-item">
-            <a th:classappend="${#ctx.springRequestContext.requestUri.equals('/pointurl')} ? 'active' : ''" class="nav-link" href="/pointurl">
+            <a th:classappend="${#ctx.springRequestContext.requestUri.equals('/pointUrl')} ? 'active' : ''" class="nav-link" href="/pointUrl">
               <span data-feather="bar-chart-2" class="align-text-bottom"></span>
               포인트 URL
             </a>
